moved map feature code to mapUtil and fixed broken test
